Add files via upload
The Color Flipper is a simple yet engaging web application that allows users to change the background color of the page randomly with the click of a button. Built using HTML, CSS, and JavaScript, this project showcases a clean and modern user interface, featuring a centered layout with a stylish button and a display area for the current color code. The application utilizes a smooth transition effect for the background color change, enhancing the user experience. This project is perfect for beginners looking to practice their web development skills and create interactive web applications.