Add files via upload
Write short description for this project


This project is a Animated Hamburger Menu built with HTML, CSS, and JavaScript, designed for mobile-friendly navigation. It features a sleek, modern design with a smooth toggle animation, transforming the hamburger icon into a visible cross when opened. The fully responsive layout ensures an optimal user experience across devices, with a sliding menu on mobile and a horizontal layout on desktop. The high-quality, optimized code follows JavaScript best practices, with efficient event handling and state management, all separated into distinct files for maintainability and performance.